pyproject.toml
[project]
name = "aws-mcp-fastapi"
version = "0.1.0"
description = "aws-doc-mcp-fastapi-streamlit"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "langchain-groq>=1.1.1",
    "mcp-use>=1.5.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]
